Refactor to remove unnecessary changes
